+++ lib/jsapi/controller/parameters.rb
@@ -30,7 +30,7 @@ module Jsapi
 
         # Wrap params
         meta_models.each do |name, meta_model|
-          @raw_attributes[name] = JSON.wrap(params[name], meta_model.schema, definitions)
+          @raw_attributes[name.underscore] = JSON.wrap(params[name], meta_model.schema, definitions)
         end
       end